apps/tuya_t5_pixel/tuya_t5_pixel_bongocat_kb/pc_kb_event/requirements.txt
pynput>=1.7.6
pyserial>=3.5

